+++ src/type_PauliBasis.jl
@@ -130,7 +130,7 @@ Base.adjoint(p::PauliBasis) = p
 function Base.iterate(::Type{PauliBasis{N}}, state = 1) where N
     state > 4^N && return
     next = CartesianIndices((2^N,2^N))[state]
-    return PauliBasis{N}(next[1]-1, next[2]-1), state+1 
+    return PauliBasis{N}(Int128(next[1]-1), Int128(next[2]-1)), state+1
 end
  
 @inline commute(p1::PauliBasis, p2::PauliBasis) = iseven(count_ones(p1.x & p2.z) - count_ones(p1.z & p2.x))
